requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
langchain
chromadb
streamlit
python-dotenv
watchdog
sentence-transformers
